--- outputs/on_mars/pr/adj_rd_s10_e8_pr.csv
@@ -1,9 +1,0 @@
-adj_rd_s10_e8,pr,790.915,6777.96,6155.48,6160.72,6145.24,6250.31,6148.44,6152.18,6146.21,6158.49,6162.35
-adj_rd_s10_e8,pr,797.697,6838.88,6241.45,6232.52,6216.96,6207.48,6276.28,6220.13,6211.62,6227.86,6213.12
-adj_rd_s10_e8,pr,793.627,6846.64,6221.2,6226.65,6208.02,6212.26,6216.8,6221.53,6213.51,6201.19,6219.73
-adj_rd_s10_e8,pr,800.547,6746.88,6128.11,6175.06,6111.44,6118.7,6111.34,6121.49,6111.66,6119.46,6116.47
-adj_rd_s10_e8,pr,796.586,6769.63,6165.25,6139.79,6173.46,6151.62,6143.91,6153.78,6143.42,6183.03,6160
-adj_rd_s10_e8,pr,795.613,6776.7,6149.99,6126.65,6134.87,6122.78,6138.19,6125.26,6137.91,6124.54,6131.09
-adj_rd_s10_e8,pr,795.127,6761.56,6138.99,6136.39,6121.48,6131.24,6118.45,6123.28,6113.62,6124.71,6144.69
-adj_rd_s10_e8,pr,800.786,6819.24,6197.48,6199.32,6199.56,6198.22,6255,6181.16,6209.97,6189.52,6196.18
-adj_rd_s10_e8,pr,863.054,7914.1,7217.3,7191.21,7200.75,7210.72,7209.18,7193.73,7175.06,7198.53,7196.4
