--- outputs/on_mars/pr/cit-Patents.txt
@@ -1,1 +1,0 @@
-
