--- outputs/on_mars/pr/ekey_rd_s10_e8_pr.csv
@@ -1,9 +1,0 @@
-ekey_rd_s10_e8,pr,4711.67,28523,26604.7,26660.5,26571,26610.6,26683.7,26684.1,26684.5,26721.8,26746.5
-ekey_rd_s10_e8,pr,4510.9,27974.1,27664.3,27805.4,27694.6,27709,27629.6,27737.8,27763.4,27704,27685.8
-ekey_rd_s10_e8,pr,4317.69,26929.5,26508.2,26647.9,26554.3,26521.8,26524.2,26572.5,26472.2,26553.6,26499.8
-ekey_rd_s10_e8,pr,4543.13,28070.1,27906.7,27809.4,27930.8,28753.5,30508.4,30354.2,30422.8,30439.2,30379.6
-ekey_rd_s10_e8,pr,4904.6,31582.7,31256.3,31301.6,31451.2,31193.9,31454.9,31198.2,31318.5,31392,31466.4
-ekey_rd_s10_e8,pr,4487.35,28332.9,28031.8,28040,28093.9,28011.5,28025.6,28020.6,28046.2,55021.9,56166.4
-ekey_rd_s10_e8,pr,4558.86,29029.8,28829.9,28817.9,28829.5,28797.1,28738.3,28820.6,28771.8,28790.2,28726.8
-ekey_rd_s10_e8,pr,4718.17,29510.9,29455.8,29228.4,29246.7,29359.8,29307.4,29338.1,29333.8,29258.8,29190.4
-ekey_rd_s10_e8,pr,4973.01,31503.5,31389.6,31421.4,31386.2,31405,31386.2,31337.4,31507.6,31411,31448.7
