--- outputs/on_mars/pr/std_rd_s10_e8_pr.csv
@@ -1,9 +1,0 @@
-std_rd_s10_e8,pr,801.052,27961.2,26573.3,26679.2,26532.9,26553.3,26528,26538.7,26486.9,26512.7,26521
-std_rd_s10_e8,pr,817.91,28641,27233.5,27183.3,27148.2,27424.3,27445.8,27435.5,27439,27200.8,27268.6
-std_rd_s10_e8,pr,810.195,27289.4,26022.8,25951.6,25919.2,26133.4,26062.4,26064.3,26155.7,25814.5,25846.1
-std_rd_s10_e8,pr,821.383,28797,26108,26129.4,26053.4,26140.5,28447.8,28473.9,28460.4,28275.7,28265.1
-std_rd_s10_e8,pr,810.453,27877.6,26522,26549.5,26448.1,26560.4,26571.7,26543.5,26548,26436.6,26415.8
-std_rd_s10_e8,pr,797.626,26985,25696.2,25858,25596.1,25603.5,25501.5,25545.3,25540.6,25698,25676.4
-std_rd_s10_e8,pr,801.422,27648.4,26298.3,26270.5,26184,26352.4,26351.2,26317.3,26412.3,26219.6,26234.6
-std_rd_s10_e8,pr,810.789,27592.9,26206.5,26153.1,26146,26131.3,26239.4,26130.6,26131.9,26122.8,26147.4
-std_rd_s10_e8,pr,844.937,30176.4,28792.2,28795.8,28776.7,28709.4,28733.2,28677.3,28671.8,28776.5,28820.5
